[feat](router.js, stations.js, ...) add routing
This is a particularly big commit because I had to add the iron:router package and make new template files so that it is routing to the proper templates.
I also commented out the subscribe in main.js (will probably delete the file later) as the subscription is handled via the router which waits for the stations subscription to become available before loading the template.